Rename Message and move package
Message 클래스의 이름을 RequestMessage 로 변경하고

websocket.vo 패키지를 handler.vo 로 이동
 - request / response 는 Handler/Dispatcher 에서 메시지 핸들링을 위하여 사용되는 파라미터이다.
 - SlackMessageHandler 는 Slack 으로 부터 받은 Raw 메시지를
   Handler 와 Dispatcher 가 이해할 수 있는 Request 객체로 변환하는 책임만 있다.
