+++ src/main/java/lets/trasnfer/bot/handler/MessageHandler.java
@@ -1,8 +1,8 @@
 package lets.trasnfer.bot.handler;
 
-import lets.trasnfer.bot.websocket.vo.Message;
-import lets.trasnfer.bot.websocket.vo.ResponseMessage;
+import lets.trasnfer.bot.handler.vo.RequestMessage;
+import lets.trasnfer.bot.handler.vo.ResponseMessage;
 
 public interface MessageHandler {
-	ResponseMessage handle(Message message);
+	ResponseMessage handle(RequestMessage message);
 }
