+++ src/main/java/lets/trasnfer/bot/handler/MovieInfoHandler.java
@@ -2,22 +2,20 @@ package lets.trasnfer.bot.handler;
 
 import com.squareup.okhttp.OkHttpClient;
 import lets.trasnfer.bot.handler.movie.MovieResponse;
-import lets.trasnfer.bot.websocket.vo.Message;
-import lets.trasnfer.bot.websocket.vo.ResponseMessage;
+import lets.trasnfer.bot.handler.vo.RequestMessage;
+import lets.trasnfer.bot.handler.vo.ResponseMessage;
 import org.springframework.http.client.ClientHttpRequestFactory;
 import org.springframework.http.client.OkHttpClientHttpRequestFactory;
-import org.springframework.http.converter.xml.MarshallingHttpMessageConverter;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponentsBuilder;
 import java.net.URI;
-import java.util.Arrays;
 
 public class MovieInfoHandler implements MessageHandler {
     private final String apiHost = "apis.daum.net";
     private String apiKey = "abcd";
 
     @Override
-    public ResponseMessage handle(Message message) {
+    public ResponseMessage handle(RequestMessage message) {
         String[] split = message.getText().split(" ");
 
         OkHttpClient client = new OkHttpClient();
