+++ src/main/java/lets/trasnfer/bot/handler/WeatherHandler.java
@@ -1,11 +1,11 @@
 package lets.trasnfer.bot.handler;
 
-import lets.trasnfer.bot.websocket.vo.Message;
-import lets.trasnfer.bot.websocket.vo.ResponseMessage;
+import lets.trasnfer.bot.handler.vo.RequestMessage;
+import lets.trasnfer.bot.handler.vo.ResponseMessage;
 
 public class WeatherHandler implements MessageHandler {
 	@Override
-	public ResponseMessage handle(Message message) {
+	public ResponseMessage handle(RequestMessage message) {
 		// 1. api를 통해서 날씨 정보 가져온다
 		// 2. 가져온 정보를 알맞게 parsing해서 리턴
