+++ src/main/java/lets/trasnfer/bot/websocket/SlackMessageHandler.java
@@ -3,8 +3,8 @@ package lets.trasnfer.bot.websocket;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lets.trasnfer.bot.handler.MessageDispatcher;
-import lets.trasnfer.bot.websocket.vo.Message;
-import lets.trasnfer.bot.websocket.vo.ResponseMessage;
+import lets.trasnfer.bot.handler.vo.RequestMessage;
+import lets.trasnfer.bot.handler.vo.ResponseMessage;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.socket.CloseStatus;
 import org.springframework.web.socket.TextMessage;
@@ -34,10 +34,10 @@ class SlackMessageHandler implements WebSocketHandler {
 	public void handleMessage(WebSocketSession session, WebSocketMessage<?> webSocketMessage) throws Exception {
 		final String payload = webSocketMessage.getPayload().toString();
 		log.debug("Payload: {}", payload);
-		final Message message = objectMapper.readValue(payload, Message.class);
+		final RequestMessage message = objectMapper.readValue(payload, RequestMessage.class);
 		log.debug("Message : {} ", message);
 
-		if (message.ofType("message")) {
+		if (message.ofType("request")) {
 			ResponseMessage response = dispatcher.getHandleMessage(message);
 			session.sendMessage(new TextMessage(objectMapper.writeValueAsString(response)));
 		}
